V 0.03 Graze Optimization
- Modified how grazing bullets affects score.  Instead of adding one
point per graze, all other points are multiplied by 1 + (Graze/100).
Because of this, grazing acts as a multiplier for longer levels.

- Modified health bar
Instead of using a set health for the bar, the boss now has a healthbar
that grabs it's HP.

UPCOMING:
Add healthbar Rendering to the boss class so that it can be implemented
and removed more effectively.

Add "Spell cards"

Figure out how to do enemies.